+++ Sources/StreamChat/Models/Channel.swift
@@ -157,6 +157,9 @@ public struct ChatChannel {
     /// because the preview message is the last `non-deleted` message sent to the channel.
     public let previewMessage: ChatMessage?
 
+    /// The draft message in the channel.
+    public let draftMessage: DraftMessage?
+
     // MARK: - Internal
 
     var hasUnread: Bool {
@@ -194,7 +197,8 @@ public struct ChatChannel {
         lastMessageFromCurrentUser: ChatMessage?,
         pinnedMessages: [ChatMessage],
         muteDetails: MuteDetails?,
-        previewMessage: ChatMessage?
+        previewMessage: ChatMessage?,
+        draftMessage: DraftMessage?
     ) {
         self.cid = cid
         self.name = name
@@ -227,6 +231,7 @@ public struct ChatChannel {
         self.pinnedMessages = pinnedMessages
         self.muteDetails = muteDetails
         self.previewMessage = previewMessage
+        self.draftMessage = draftMessage
     }
 
     /// Returns a new `ChatChannel` with the provided data replaced.
@@ -266,7 +271,8 @@ public struct ChatChannel {
             lastMessageFromCurrentUser: lastMessageFromCurrentUser,
             pinnedMessages: pinnedMessages,
             muteDetails: muteDetails,
-            previewMessage: previewMessage
+            previewMessage: previewMessage,
+            draftMessage: draftMessage
         )
     }
 }
@@ -315,6 +321,7 @@ extension ChatChannel: Hashable {
         guard lhs.team == rhs.team else { return false }
         guard lhs.truncatedAt == rhs.truncatedAt else { return false }
         guard lhs.ownCapabilities == rhs.ownCapabilities else { return false }
+        guard lhs.draftMessage == rhs.draftMessage else { return false }
         return true
     }