+++ Sources/StreamChat/Models/ChatMessage.swift
@@ -64,6 +64,9 @@ public struct ChatMessage {
     public var quotedMessage: ChatMessage? { _quotedMessage() }
     let _quotedMessage: () -> ChatMessage?
 
+    /// The draft reply to this message. Applies only for the messages of the current user.
+    public let draftReply: DraftMessage?
+
     /// A flag indicating whether the message was bounced due to moderation.
     public let isBounced: Bool
 
@@ -213,7 +216,8 @@ public struct ChatMessage {
         moderationDetails: MessageModerationDetails?,
         readBy: Set<ChatUser>,
         poll: Poll?,
-        textUpdatedAt: Date?
+        textUpdatedAt: Date?,
+        draftReply: DraftMessage?
     ) {
         self.id = id
         self.cid = cid
@@ -254,6 +258,7 @@ public struct ChatMessage {
         self.readBy = readBy
         _attachments = attachments
         _quotedMessage = { quotedMessage }
+        self.draftReply = draftReply
     }
 
     /// Returns a new `ChatMessage` with the provided data replaced.
@@ -300,7 +305,8 @@ public struct ChatMessage {
             moderationDetails: moderationDetails,
             readBy: readBy,
             poll: poll,
-            textUpdatedAt: textUpdatedAt
+            textUpdatedAt: textUpdatedAt,
+            draftReply: draftReply
         )
     }
 }
@@ -431,6 +437,7 @@ extension ChatMessage: Hashable {
         guard lhs.quotedMessage == rhs.quotedMessage else { return false }
         guard lhs.translations == rhs.translations else { return false }
         guard lhs.type == rhs.type else { return false }
+        guard lhs.draftReply == rhs.draftReply else { return false }
         return true
     }